Switch to PNG tiles to retrieve the ground altitude.
The previous skadi tiles were 1x1degree and gzipped.
Each tile is ~10MB gzipped and expands to ~25MB.
Caching those tiles was causing an OOM on Cloud Run.

This PR switches to the 256x256px PNG tiles which are ~0.2MB each.
The download size and time should be vastly improved.

The LRU cache is configured in MB.

Fixes #37.