Make live track 3d markers appear on the ground.
This is mostly to prevent messages to be displayed under the ground. It uses an ElevationSampler to query the ground altitude. Note that sampler is updated as the globe moves while the marker elevation is only updated on refreshes. It might cause a few visual glitches for pilots and messages close to the ground after the globe is moved and before it is refreshed.